+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Updates splashscreen for both platforms
 * Fixes app title for ios
 * Updates the marker for the Solar Car, so it looks better
+* Updates the team details page with information about every team member.
 
 ## 0.0.2